improved search performance
